+++ client/package.json
@@ -44,6 +44,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-    "proxy" : "http://localhost:8000"
+  }
 }
